New solver: "bounded" succinct small progress measures (BSSPM)
Rewrite of the SSPM solver to use the same code for the even and for the
odd measures. Added a variation of SSPM that starts with a low value for
the number of bits in the adaptive counters, for now called the bounded
succinct small progress measures, as we start with a low bound for the
number of bits which the increases when the bound is too low. The idea
behind this variation comes from a discussion with Marcin Jurdzinski.