New version of distraction-free tangle learning.
This version is not 'as exponential' as the previous attempt.
Every iteration partitions the game into smaller blocks, disjunctively.
We perform single-player tangle learning in some blocks.
Each iteration costs time polynomial in the number of tangles and the size of the game.
The number of iterations determines the runtime complexity of this algorithm.